requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0
django-ckeditor==5.4.0
django-js-asset==1.0.0
django-simditor==0.0.10
idna==2.6
Markdown==2.6.11
oauthlib==2.1.0
Pillow==5.1.0
Pygments==2.2.0
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
